Fixed assimp import issue.

+++ src/skeletal_mesh.h
@@ -26,9 +26,9 @@ struct SkeletalColoredVertex
 
 struct SubMesh
 {
-	uint16_t num_indices = 0;
-	uint16_t base_vertex = 0;
-	uint16_t base_index = 0;
+	uint32_t num_indices = 0;
+	uint32_t base_vertex = 0;
+	uint32_t base_index = 0;
 };
 
 class SkeletalMesh
